feat(groups): Change routes for fetching groups
