+++ config/routes.rb
@@ -23,7 +23,8 @@ Rails.application.routes.draw do
 
       resources :add_ons, param: :code
       resources :billable_metrics, param: :code
-      resources :groups, only: :index
+      get '/billable_metrics/:code/groups', to: 'billable_metrics/groups#index', as: 'billable_metric_groups'
+
       resources :coupons, param: :code
       resources :credit_notes, only: %i[show index] do
         post :download, on: :member
